# SERVER_STABLE/SERVER_STABLE/openapi.yaml
openapi: 3.0.3
info:
  title: Academy Portal API
  version: 1.0.0
  description: >
    Technical contract for the Academy Portal backend. The API supports public
    course discovery, student enrollments, and admin management of courses,
    teachers, and categories.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/users/me:
    get:
      tags:
        - Users
      summary: Return the current authenticated user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Users
      summary: Update the current authenticated user's profile
      operationId: updateCurrentUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
            example:
              display_name: Ayan Sarsen
      responses:
        '200':
          description: Profile updated and confirmation email queued when changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/me/password:
    put:
      tags:
        - Users
      summary: Change the current authenticated user's password
      operationId: changeCurrentUserPassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
            example:
              current_password: StrongPass123
              new_password: NewStrongPass123
      responses:
        '200':
          description: Password changed and confirmation email queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/teachers:
    get:
      tags:
        - Teachers
      summary: List teachers
      operationId: listTeachers
      security: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          example: aigerim
        - name: subject
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          example: Web Development
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - name
              - -created_at
              - created_at
          example: name
      responses:
        '200':
          description: Paginated teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Teachers
      summary: Create a teacher profile
      operationId: createTeacher
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherCreateRequest'
      responses:
        '201':
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/teachers/{teacherId}:
    put:
      tags:
        - Teachers
      summary: Update a teacher profile
      operationId: updateTeacher
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeacherIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherUpdateRequest'
      responses:
        '200':
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Teachers
      summary: Delete a teacher profile
      operationId: deleteTeacher
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeacherIdParam'
      responses:
        '200':
          description: Teacher deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: >
        Opaque keyset cursor generated from the previous page using
        (created_at, id) or (enrolled_at, id) ordering.
      schema:
        type: string
      example: MjAyNi0wNC0yOFQwOTowMDowMFouMDAwWjo5MDAx
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    TeacherIdParam:
      name: teacherId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
      example: '3001'
  responses:
    BadRequest:
      description: Request body, path, or query parameter is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: bad_request
              message: Validation failed for query parameter limit.
              details:
                limit: must be less than or equal to 100
              request_id: req_7f3a0d2e
    Unauthorized:
      description: Missing, expired, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Bearer token is missing or invalid.
              details: {}
              request_id: req_7f3a0d2e
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: You do not have permission to perform this action.
              details:
                required_roles:
                  - admin
              request_id: req_7f3a0d2e
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: Course 9001 was not found.
              details: {}
              request_id: req_7f3a0d2e
    Conflict:
      description: Request conflicts with current system state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: conflict
              message: Student is already enrolled in this course.
              details:
                course_id: '9001'
              request_id: req_7f3a0d2e
    UnprocessableEntity:
      description: Request format is valid JSON but violates semantic validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unprocessable_entity
              message: Course cannot be published while required fields are missing.
              details:
                missing_fields:
                  - description
                  - capacity
              request_id: req_7f3a0d2e
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_server_error
              message: An unexpected error occurred.
              details: {}
              request_id: req_7f3a0d2e
  schemas:
    Id:
      type: string
      pattern: '^[0-9]+$'
      example: '9001'
...
    Role:
      type: string
      enum:
        - student
        - admin
      example: student
...
    Error:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          example: conflict
        message:
          type: string
          example: Student is already enrolled in this course.
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
          example: req_7f3a0d2e
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    MutationSuccess:
      type: object
      required:
        - ok
        - message
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Resource deleted successfully.
    PaginationMeta:
      type: object
      required:
        - limit
        - has_more
      properties:
        next_cursor:
          type: string
          nullable: true
          example: MjAyNi0wNC0yN1QxMTowMDowMFouMDAwWjo4OTk1
        limit:
          type: integer
          example: 20
        has_more:
          type: boolean
          example: true
    UserSummary:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          format: email
          example: student@academy.dev
        display_name:
          type: string
          nullable: true
          example: Ayan Sarsen
        role:
          $ref: '#/components/schemas/Role'
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-05T10:05:00Z'
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-04-28T09:00:00Z'
...
    EmailJobSummary:
      type: object
      properties:
        queued:
          type: boolean
          example: true
        job_id:
          type: string
          nullable: true
          example: '42'
        fallback:
          type: string
          nullable: true
          example: setImmediate
...
    MeResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/UserSummary'
    UserMutationResponse:
      type: object
      required:
        - ok
        - message
        - data
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Account profile updated. A confirmation email has been queued.
        data:
          $ref: '#/components/schemas/UserSummary'
        email_job:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/EmailJobSummary'
    UserProfileUpdateRequest:
      type: object
      properties:
        display_name:
          type: string
          nullable: true
          maxLength: 120
          example: Ayan Sarsen
    PasswordChangeRequest:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          maxLength: 4096
          example: StrongPass123
        new_password:
          type: string
          minLength: 8
          example: NewStrongPass123
...
    Teacher:
      type: object
      required:
        - id
        - full_name
        - subject
        - rating
        - is_active
      properties:
        id:
          $ref: '#/components/schemas/Id'
        full_name:
          type: string
          example: Aigerim Tolegen
        subject:
          type: string
          example: Web Development
        rating:
          type: number
          format: float
          example: 4.9
        bio:
          type: string
          nullable: true
          example: Senior frontend engineer with 8 years of product experience.
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-04-10T09:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-04-25T09:00:00Z'
    TeacherCreateRequest:
      type: object
      required:
        - full_name
        - subject
        - rating
      properties:
        full_name:
          type: string
          example: Aigerim Tolegen
        subject:
          type: string
          example: Web Development
        rating:
          type: number
          format: float
          minimum: 1
          maximum: 5
          example: 4.9
        bio:
          type: string
          nullable: true
          example: Senior frontend engineer with 8 years of product experience.
        is_active:
          type: boolean
          default: true
          example: true
    TeacherUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/TeacherCreateRequest'
    TeacherListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...